pyproject.toml
[project]
name = "media-mporter"
version = "0.1.0"
description = "Cross-platform media importer from camera SD cards"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "SQLAlchemy>=2.0",
    "Pillow>=10.0",
    "exifread>=3.0",
    "pymediainfo>=6.0",
    "watchdog>=3.0",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
